refactor: getCssSelector함수 구분
